+++ engine/engine/core/Image.cpp
@@ -52,7 +52,7 @@ void Image::Load( S<Image> img, const Blob& binary )
 {
    int w,h;
    int channelCount;
-   unsigned char* imageData = stbi_load_from_memory((const stbi_uc*)binary.Data(), binary.Size(), &w, &h, &channelCount, 4);
+   unsigned char* imageData = stbi_load_from_memory((const stbi_uc*)binary.Data(), (int)binary.Size(), &w, &h, &channelCount, 4);
    img.reset(new Image(w, h, eTextureFormat::RGBA8Unorm, imageData, w * h * 4));
    stbi_image_free( imageData );
 }
@@ -62,7 +62,7 @@ bool Asset<Image>::Load(S<Image>& res, const void* binary, size_t size)
 {
 	int w, h;
 	int channelCount;
-	unsigned char* imageData = stbi_load_from_memory((const stbi_uc*)binary, size, &w, &h, &channelCount, 4);
+	unsigned char* imageData = stbi_load_from_memory((const stbi_uc*)binary, (int)size, &w, &h, &channelCount, 4);
 	res.reset(new Image(w, h, eTextureFormat::RGBA8Unorm, imageData, w * h * 4));
 	stbi_image_free(imageData);
 	return true;
